Implemented and tested the GPT2 version of the architcture, and working HLS.

+++ Model-architectures/gpt2-HARDWARE/HLS-Verilog/Compute_Controller_Logic/src-hls/compute_controller.hpp
@@ -10,7 +10,7 @@ void compute_controller(
 
     // FSM communication signals
     bool        compute_start,       // [INPUT] Start signal for compute
-    uint32_t    compute_instruction, // [INPUT] Compute operation [7:0]=op [15:8]=layer [23:16]=head [31:24]=tile
+    uint64_t    compute_instruction, // [INPUT] Compute operation [7:0]=op [15:8]=layer [23:16]=head [55:24]=tile
     bool        &compute_ready,      // [OUTPUT] Compute engine ready for new operation
     bool        &compute_done,       // [OUTPUT] Compute operation finished
 
@@ -18,7 +18,7 @@ void compute_controller(
     bool        mem_transfer_done,   // [INPUT] Memory manager transfer complete
     bool        &mem_read_request,   // [OUTPUT] Request memory manager read
     bool        &mem_write_request,  // [OUTPUT] Request memory manager write
-    uint32_t    &mem_op,             // [OUTPUT] Opcode for memory manager
+    uint64_t    &mem_op,             // [OUTPUT] Opcode for memory manager
 
     // Flat input/output buffers
     const uint8_t in_buf[compute_buf::IN_BUF_BYTES],
@@ -26,11 +26,11 @@ void compute_controller(
 
     // Debug visibility
     ComputeState &dbg_state,
-    uint32_t    &dbg_req_instruction,
+    uint64_t    &dbg_req_instruction,
     uint8_t     &dbg_req_op,
     uint8_t     &dbg_req_layer,
     uint8_t     &dbg_req_head,
-    uint8_t     &dbg_req_tile,
+    uint16_t    &dbg_req_tile,
     bool        &dbg_mac_start,
     bool        &dbg_mac_ready,
     bool        &dbg_mac_complete,
